refactor: extract utils, hoist cwd up
